+++ packages/ui/src/search/sources/opentreeoflife.org/OTOLResolve.tsx
@@ -99,4 +99,3 @@ export const OTOLResolve: React.FC = () => {
         </>
     )
 }
-export default OTOLResolve