+++ packages/ui/src/search/sources/paleobiodb.org/PBDBResolve.tsx
@@ -101,4 +101,3 @@ export const PBDBResolve: React.FC = () => {
         </>
     )
 }
-export default PBDBResolve