+++ apps/www/CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   GBIF to search capabilities.
+
 ### Changed
+
+-   Search suggestions return full URL.
+-   Search: node results appear in a separate container from resolutions.
 
 ### Deprecated